+++ mapbox_gl_web/lib/src/convert.dart
@@ -56,11 +56,18 @@ class Convert {
           options['logoViewMargins'][0], options['logoViewMargins'][1]);
     }
     if (options.containsKey('compassViewPosition')) {
-      sink.setCompassGravity(options['compassViewPosition']);
+      final position =
+          CompassViewPosition.values[options['compassViewPosition']];
+      sink.setCompassAlignment(position);
     }
     if (options.containsKey('compassViewMargins')) {
       sink.setCompassViewMargins(
           options['compassViewMargins'][0], options['compassViewMargins'][1]);
+    }
+    if (options.containsKey('attributionButtonPosition')) {
+      final position = AttributionButtonPosition
+          .values[options['attributionButtonPosition']];
+      sink.setAttributionButtonAlignment(position);
     }
     if (options.containsKey('attributionButtonMargins')) {
       sink.setAttributionButtonMargins(options['attributionButtonMargins'][0],
@@ -203,7 +210,10 @@ class Convert {
       properties['iconRotate'] = options.iconRotate;
     }
     if (options.iconOffset != null) {
-      properties['iconOffset'] = [options.iconOffset!.dx, options.iconOffset!.dy];
+      properties['iconOffset'] = [
+        options.iconOffset!.dx,
+        options.iconOffset!.dy
+      ];
     }
     if (options.iconAnchor != null) {
       properties['iconAnchor'] = options.iconAnchor;
@@ -233,7 +243,10 @@ class Convert {
       properties['textTransform'] = options.textTransform;
     }
     if (options.textOffset != null) {
-      properties['textOffset'] = [options.textOffset!.dx, options.textOffset!.dy];
+      properties['textOffset'] = [
+        options.textOffset!.dx,
+        options.textOffset!.dy
+      ];
     }
     if (options.iconOpacity != null) {
       properties['iconOpacity'] = options.iconOpacity;