+++ mapbox_gl_web/lib/src/mapbox_map_controller.dart
@@ -1,7 +1,7 @@
 part of mapbox_gl_web;
 
 const _mapboxGlCssUrl =
-    'https://api.tiles.mapbox.com/mapbox-gl-js/v1.6.1/mapbox-gl.css';
+    'https://api.mapbox.com/mapbox-gl-js/v2.6.1/mapbox-gl.css';
 
 class MapboxMapController extends MapboxGlPlatform
     implements MapboxMapOptionsSink {
@@ -9,8 +9,10 @@ class MapboxMapController extends MapboxGlPlatform
 
   late Map<String, dynamic> _creationParams;
   late MapboxMap _map;
+  bool _mapReady = false;
 
   List<String> annotationOrder = [];
+  final _featureLayerIdentifiers = Set<String>();
   late SymbolManager symbolManager;
   late LineManager lineManager;
   late CircleManager circleManager;
@@ -63,6 +65,13 @@ class MapboxMapController extends MapboxGlPlatform
         ),
       );
       _map.on('load', _onStyleLoaded);
+      _map.on('click', _onMapClick);
+      // long click not available in web, so it is mapped to double click
+      _map.on('dblclick', _onMapLongClick);
+      _map.on('movestart', _onCameraMoveStarted);
+      _map.on('move', _onCameraMove);
+      _map.on('moveend', _onCameraIdle);
+      _map.on('resize', _onMapResize);
     }
     Convert.interpretMapboxMapOptions(_creationParams['options'], this);
 
@@ -110,7 +119,7 @@ class MapboxMapController extends MapboxGlPlatform
 
   @override
   Future<void> matchMapLanguageWithDeviceDefault() async {
-    setMapLanguage(ui.window.locale!.languageCode);
+    setMapLanguage(ui.window.locale.languageCode);
   }
 
   @override
@@ -166,6 +175,12 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
+  Future<LatLng> getSymbolLatLng(Symbol symbol) async {
+    var coordinates = symbolManager.getFeature(symbol.id)!.geometry.coordinates;
+    return LatLng(coordinates[1], coordinates[0]);
+  }
+
+  @override
   Future<void> removeSymbols(Iterable<String> symbolsIds) async {
     symbolManager.removeAll(symbolsIds);
   }
@@ -187,6 +202,13 @@ class MapboxMapController extends MapboxGlPlatform
   @override
   Future<void> updateLine(Line line, LineOptions changes) async {
     lineManager.update(line.id, changes);
+  }
+
+  @override
+  Future<List<LatLng>> getLineLatLngs(Line line) async {
+    List<dynamic> coordinates =
+        lineManager.getFeature(line.id)!.geometry.coordinates;
+    return coordinates.map((c) => LatLng(c[1], c[0])).toList();
   }
 
   @override
@@ -353,6 +375,11 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
+  Future<void> removeSource(String sourceId) {
+    return _map.removeSource(sourceId);
+  }
+
+  @override
   Future<void> setSymbolIconAllowOverlap(bool enable) async {
     //TODO: to implement
     print('setSymbolIconAllowOverlap not implemented yet');
@@ -390,6 +417,7 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   void _onStyleLoaded(_) {
+    _mapReady = true;
     for (final annotationType in annotationOrder) {
       switch (annotationType) {
         case 'AnnotationType.symbol':
@@ -411,15 +439,7 @@ class MapboxMapController extends MapboxGlPlatform
               "Unknown annotation type: \(annotationType), must be either 'fill', 'line', 'circle' or 'symbol'");
       }
     }
-
     onMapStyleLoadedPlatform(null);
-    _map.on('click', _onMapClick);
-    // long click not available in web, so it is mapped to double click
-    _map.on('dblclick', _onMapLongClick);
-    _map.on('movestart', _onCameraMoveStarted);
-    _map.on('move', _onCameraMove);
-    _map.on('moveend', _onCameraIdle);
-    _map.on('resize', _onMapResize);
   }
 
   void _onMapResize(Event e) {
@@ -434,11 +454,19 @@ class MapboxMapController extends MapboxGlPlatform
     });
   }
 
-  void _onMapClick(e) {
-    onMapClickPlatform({
-      'point': Point<double>(e.point.x, e.point.y),
-      'latLng': LatLng(e.lngLat.lat, e.lngLat.lng),
-    });
+  void _onMapClick(Event e) {
+    final features = _map.queryRenderedFeatures(
+        [e.point.x, e.point.y], {"layers": _featureLayerIdentifiers.toList()});
+    final payload = {
+      'point': Point<double>(e.point.x.toDouble(), e.point.y.toDouble()),
+      'latLng': LatLng(e.lngLat.lat.toDouble(), e.lngLat.lng.toDouble()),
+      if (features.isNotEmpty) "id": features.first.id,
+    };
+    if (features.isNotEmpty) {
+      onFeatureTappedPlatform(payload);
+    } else {
+      onMapClickPlatform(payload);
+    }
   }
 
   void _onMapLongClick(e) {
@@ -612,8 +640,13 @@ class MapboxMapController extends MapboxGlPlatform
   }
 
   @override
-  void setCompassGravity(int gravity) {
-    _updateNavigationControl(position: CompassViewPosition.values[gravity]);
+  void setCompassAlignment(CompassViewPosition position) {
+    _updateNavigationControl(position: position);
+  }
+
+  @override
+  void setAttributionButtonAlignment(AttributionButtonPosition position) {
+    print('setAttributionButtonAlignment not available in web');
   }
 
   @override
@@ -687,7 +720,19 @@ class MapboxMapController extends MapboxGlPlatform
 
   @override
   void setStyleString(String? styleString) {
+    //remove old mouseenter callbacks to avoid multicalling
+    for (var layerId in _featureLayerIdentifiers) {
+      _map.off('mouseenter', layerId, _onMouseEnterFeature);
+      _map.off('mousemouve', layerId, _onMouseEnterFeature);
+      _map.off('mouseleave', layerId, _onMouseLeaveFeature);
+    }
+    _featureLayerIdentifiers.clear();
+
     _map.setStyle(styleString);
+    // catch style loaded for later style changes
+    if (_mapReady) {
+      _map.once("styledata", _onStyleLoaded);
+    }
   }
 
   @override
@@ -752,5 +797,102 @@ class MapboxMapController extends MapboxGlPlatform
     var circumference = 40075017.686;
     var zoom = _map.getZoom();
     return circumference * cos(latitude * (pi / 180)) / pow(2, zoom + 9);
+  }
+
+  @override
+  Future<void> removeLayer(String layerId) async {
+    _featureLayerIdentifiers.remove(layerId);
+    _map.removeLayer(layerId);
+  }
+
+  @override
+  Future<void> addGeoJsonSource(String sourceId, Map<String, dynamic> geojson,
+      {String? promoteId}) async {
+    _map.addSource(sourceId, {
+      "type": 'geojson',
+      "data": geojson,
+      if (promoteId != null) "promoteId": promoteId
+    });
+  }
+
+  @override
+  Future<void> setGeoJsonSource(
+      String sourceId, Map<String, dynamic> geojson) async {
+    final source = _map.getSource(sourceId) as GeoJsonSource;
+    final data = FeatureCollection(features: [
+      for (final f in geojson["features"] ?? [])
+        Feature(
+            geometry: Geometry(
+                type: f["geometry"]["type"],
+                coordinates: f["geometry"]["coordinates"]),
+            properties: f["properties"],
+            id: f["id"])
+    ]);
+    source.setData(data);
+  }
+
+  @override
+  Future<void> addCircleLayer(
+      String sourceId, String layerId, Map<String, dynamic> properties,
+      {String? belowLayerId}) async {
+    return _addLayer(sourceId, layerId, properties, "circle",
+        belowLayerId: belowLayerId);
+  }
+
+  @override
+  Future<void> addFillLayer(
+      String sourceId, String layerId, Map<String, dynamic> properties,
+      {String? belowLayerId}) async {
+    return _addLayer(sourceId, layerId, properties, "fill",
+        belowLayerId: belowLayerId);
+  }
+
+  @override
+  Future<void> addLineLayer(
+      String sourceId, String layerId, Map<String, dynamic> properties,
+      {String? belowLayerId}) async {
+    return _addLayer(sourceId, layerId, properties, "line",
+        belowLayerId: belowLayerId);
+  }
+
+  @override
+  Future<void> addSymbolLayer(
+      String sourceId, String layerId, Map<String, dynamic> properties,
+      {String? belowLayerId}) async {
+    return _addLayer(sourceId, layerId, properties, "symbol",
+        belowLayerId: belowLayerId);
+  }
+
+  Future<void> _addLayer(String sourceId, String layerId,
+      Map<String, dynamic> properties, String layerType,
+      {String? belowLayerId}) async {
+    final layout = Map.fromEntries(
+        properties.entries.where((entry) => isLayoutProperty(entry.key)));
+    final paint = Map.fromEntries(
+        properties.entries.where((entry) => !isLayoutProperty(entry.key)));
+
+    _map.addLayer({
+      'id': layerId,
+      'type': layerType,
+      'source': sourceId,
+      'layout': layout,
+      'paint': paint
+    }, belowLayerId);
+
+    _featureLayerIdentifiers.add(layerId);
+    if (layerType == "fill") {
+      _map.on('mousemove', layerId, _onMouseEnterFeature);
+    } else {
+      _map.on('mouseenter', layerId, _onMouseEnterFeature);
+    }
+    _map.on('mouseleave', layerId, _onMouseLeaveFeature);
+  }
+
+  void _onMouseEnterFeature(_) {
+    _map.getCanvas().style.cursor = 'pointer';
+  }
+
+  void _onMouseLeaveFeature(_) {
+    _map.getCanvas().style.cursor = '';
   }
 }